no andan un par de cositas jeje

+++ Kernel/include/pipes.h
@@ -18,6 +18,6 @@ int16_t closePipe(int64_t fd);
 
 int64_t readPipe(int64_t fd, char * buffer, int64_t size);
 
-uint16_t writePipe(uint16_t fd, char * buffer, uint16_t size);
+int64_t writePipe(int64_t fd, char * buffer, int64_t size);
 
 #endif
